// src/house/house.interface.ts
export interface HouseInterface {
    id: number;
    name: string;
    price: string;
    type: string;
    city: string;
    address: string;
    furnitureType: string;
    bedroom: number;
    toilet: number;
    diningroom: number;
    image: Array<String>;
    detail: string;
    userId: number;
}